+++ HTGSCudaBackend/HTGSCudaBackend/rasterization/include/alpha_blend_global_ordering/buffer_utils.h
@@ -33,7 +33,7 @@ namespace htgs::rasterization::alpha_blend_global_ordering {
         float4* rgba = nullptr;
         bool* rgb_clamp_info = nullptr;
 
-        static PerPrimitiveBuffers from_blob(char*& blob, size_t n_primitives, bool store_rgba, bool store_rgb_clamp_info) {
+        static PerPrimitiveBuffers from_blob(char*& blob, int n_primitives, bool store_rgba, bool store_rgb_clamp_info) {
             PerPrimitiveBuffers buffers;
             obtain(blob, buffers.n_touched_tiles, n_primitives, 128);
             obtain(blob, buffers.offset, n_primitives, 128);
@@ -60,7 +60,7 @@ namespace htgs::rasterization::alpha_blend_global_ordering {
         cub::DoubleBuffer<uint64_t> keys;
         cub::DoubleBuffer<uint> primitive_indices;
     
-        static PerInstanceBuffers from_blob(char*& blob, size_t n_instances, int end_bit) {
+        static PerInstanceBuffers from_blob(char*& blob, int n_instances, int end_bit) {
             PerInstanceBuffers buffers;
             uint64_t* keys_current;
             obtain(blob, keys_current, n_instances, 128);
@@ -86,7 +86,7 @@ namespace htgs::rasterization::alpha_blend_global_ordering {
     struct PerTileBuffers {
         uint2* instance_ranges;
     
-        static PerTileBuffers from_blob(char*& blob, size_t n_tiles) {
+        static PerTileBuffers from_blob(char*& blob, int n_tiles) {
             PerTileBuffers buffers;
             obtain(blob, buffers.instance_ranges, n_tiles, 128);
             return buffers;
