+++ HTGSCudaBackend/HTGSCudaBackend/rasterization/include/hybrid_blend/buffer_utils.h
@@ -32,7 +32,7 @@ namespace htgs::rasterization::hybrid_blend {
         float4* rgba = nullptr;
         bool* rgb_clamp_info = nullptr;
 
-        static PerPrimitiveBuffers from_blob(char*& blob, size_t n_primitives, bool store_rgba, bool store_rgb_clamp_info) {
+        static PerPrimitiveBuffers from_blob(char*& blob, int n_primitives, bool store_rgba, bool store_rgb_clamp_info) {
             PerPrimitiveBuffers buffers;
             obtain(blob, buffers.n_touched_tiles, n_primitives, 128);
             obtain(blob, buffers.offset, n_primitives, 128);
@@ -60,7 +60,7 @@ namespace htgs::rasterization::hybrid_blend {
         cub::DoubleBuffer<KeyT> keys;
         cub::DoubleBuffer<uint> primitive_indices;
     
-        static PerInstanceBuffers from_blob(char*& blob, size_t n_instances, int end_bit) {
+        static PerInstanceBuffers from_blob(char*& blob, int n_instances, int end_bit) {
             PerInstanceBuffers buffers;
             KeyT* keys_current;
             obtain(blob, keys_current, n_instances, 128);
@@ -86,7 +86,7 @@ namespace htgs::rasterization::hybrid_blend {
     struct PerTileBuffers {
         uint2* instance_ranges;
     
-        static PerTileBuffers from_blob(char*& blob, size_t n_tiles) {
+        static PerTileBuffers from_blob(char*& blob, int n_tiles) {
             PerTileBuffers buffers;
             obtain(blob, buffers.instance_ranges, n_tiles, 128);
             return buffers;
@@ -96,9 +96,9 @@ namespace htgs::rasterization::hybrid_blend {
     struct PerPixelBuffers {
         uint* primitive_indices_core;
         float4* grad_info_core;
-        float* grad_info_tail; // rgb_tail, alpha_sum_rcp_tail, transmitance_tail, transmitance_core
+        float* grad_info_tail; // rgb_tail, alpha_sum_rcp_tail, transmittance_tail, transmittance_core
 
-        static PerPixelBuffers from_blob(char*& blob, size_t n_pixels, int K) {
+        static PerPixelBuffers from_blob(char*& blob, int n_pixels, int K) {
             PerPixelBuffers buffers;
             obtain(blob, buffers.primitive_indices_core, n_pixels * K, 128);
             obtain(blob, buffers.grad_info_core, n_pixels * K, 128);
