+++ HTGSCudaBackend/HTGSCudaBackend/rasterization/include/oit_blend/buffer_utils.h
@@ -32,7 +32,7 @@ namespace htgs::rasterization::oit_blend {
         float4* rgba = nullptr;
         bool* rgb_clamp_info = nullptr;
 
-        static PerPrimitiveBuffers from_blob(char*& blob, size_t n_primitives, bool store_MT3, bool store_rgba, bool store_rgb_clamp_info) {
+        static PerPrimitiveBuffers from_blob(char*& blob, int n_primitives, bool store_MT3, bool store_rgba, bool store_rgb_clamp_info) {
             PerPrimitiveBuffers buffers;
             obtain(blob, buffers.n_touched_tiles, n_primitives, 128);
             obtain(blob, buffers.offset, n_primitives, 128);
@@ -60,7 +60,7 @@ namespace htgs::rasterization::oit_blend {
         cub::DoubleBuffer<KeyT> keys;
         cub::DoubleBuffer<uint> primitive_indices;
     
-        static PerInstanceBuffers from_blob(char*& blob, size_t n_instances, int end_bit) {
+        static PerInstanceBuffers from_blob(char*& blob, int n_instances, int end_bit) {
             PerInstanceBuffers buffers;
             KeyT* keys_current;
             obtain(blob, keys_current, n_instances, 128);
@@ -86,7 +86,7 @@ namespace htgs::rasterization::oit_blend {
     struct PerTileBuffers {
         uint2* instance_ranges;
     
-        static PerTileBuffers from_blob(char*& blob, size_t n_tiles) {
+        static PerTileBuffers from_blob(char*& blob, int n_tiles) {
             PerTileBuffers buffers;
             obtain(blob, buffers.instance_ranges, n_tiles, 128);
             return buffers;
@@ -94,9 +94,9 @@ namespace htgs::rasterization::oit_blend {
     };
 
     struct PerPixelBuffers {
-        float* grad_info; // rgb, alpha_sum_rcp, transmitance
+        float* grad_info; // rgb, alpha_sum_rcp, transmittance
 
-        static PerPixelBuffers from_blob(char*& blob, size_t n_pixels) {
+        static PerPixelBuffers from_blob(char*& blob, int n_pixels) {
             PerPixelBuffers buffers;
             obtain(blob, buffers.grad_info, n_pixels * 5, 128);
             return buffers;
